refactor: move SpinalSimStreamWidthAdapterTester
